Add private controller. Inheritance! Projects dropdown

+++ app/controllers/memberships_controller.rb
@@ -1,4 +1,4 @@
-class MembershipsController < ApplicationController
+class MembershipsController < PrivateController
   before_action :find_set_project
 
   def index
@@ -32,13 +32,19 @@ class MembershipsController < ApplicationController
   end
 
   def destroy
-    membership = Membership.find(params[:id])
-    membership.destroy
-    flash[:notice] = "#{membership.user.full_name} was successfully removed"
-    redirect_to project_memberships_path(@project.id)
+    if owner_count(@project) == 1 && @membership.role == 1
+      flash[:danger] = "Projects must have at least one owner"
+      redirect_to project_memberships_path(@project.id)
+    else
+      membership = Membership.find(params[:id])
+      membership.destroy
+      flash[:notice] = "#{membership.user.full_name} was successfully removed"
+      redirect_to project_memberships_path(@project.id)
+    end
   end
 
   private
+
 
   def find_set_project
     @project = Project.find(params[:project_id])
